refactor: change mixins to work with modlauncher
