+++ build.gradle
@@ -5,9 +5,7 @@ plugins {
 
 repositories {
     mavenLocal()
-    maven {
-        url = "http://repo.maven.apache.org/maven2"
-    }
+    mavenCentral()
     maven {
         url = "https://libraries.minecraft.net/"
     }
@@ -20,7 +18,7 @@ dependencies {
     compile "org.projectlombok:lombok:1.18.10"
     annotationProcessor "org.projectlombok:lombok:1.18.10"
     compile "net.minecraft:server:1.14.4"
-    compile "org.spongepowered:mixin:0.7.8-SNAPSHOT"
+    compile "org.spongepowered:mixin:0.8-SNAPSHOT"
 
     compile "org.reflections:reflections:0.9.11"
     compile "com.fasterxml.jackson.core:jackson-databind:2.10.0"
@@ -30,6 +28,7 @@ dependencies {
 group = theGroup
 version = theVersion
 sourceCompatibility = "1.8"
+targetCompatibility = "1.8"
 
 tasks.withType(JavaCompile) {
     options.encoding = "UTF-8"
@@ -38,9 +37,7 @@ tasks.withType(JavaCompile) {
 jar {
     manifest {
         attributes(
-                "TweakClass" : "org.spongepowered.asm.launch.MixinTweaker",
-                "TweakOrder" : "0",
-                "MixinConfigs" : "mixins.conduit.json"
+                "MixinConnector" : "systems.conduit.main.mixin.MixinConnector"
         )
     }
 }
