+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,5 @@
 #Fri Oct 18 22:20:34 EDT 2019
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.0-rc-1-all.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.0-all.zip
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStorePath=wrapper/dists
