+++ src/main/java/systems/conduit/main/api/managers/LevelManager.java
@@ -3,16 +3,12 @@ package systems.conduit.main.api.managers;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
-import net.minecraft.server.level.progress.ChunkProgressListener;
-import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.LevelSettings;
-import net.minecraft.world.level.chunk.ChunkStatus;
 import net.minecraft.world.level.dimension.DimensionType;
 import net.minecraft.world.level.storage.LevelData;
 import net.minecraft.world.level.storage.LevelStorage;
 import systems.conduit.main.Conduit;
 
-import javax.annotation.Nullable;
 import java.util.Optional;
 
 public class LevelManager {
@@ -29,6 +25,7 @@ public class LevelManager {
 
             data.setLevelName(name);
             LevelSettings settings = new LevelSettings(data);
+            /**
             ServerLevel level = new ServerLevel((MinecraftServer) server, server.getExecutor(), storage, data, dimension, server.getProfiler(), new ChunkProgressListener() {
                 @Override
                 public void updateSpawnPos(ChunkPos chunkPos) {
@@ -48,6 +45,7 @@ public class LevelManager {
                 data.setInitialized(true);
             }
             server.getLevels().put(dimension, level);
+             */
 
             return null;
         });
