+++ src/main/java/systems/conduit/main/mixin/MixinConnector.java
@@ -1,0 +1,12 @@
+package systems.conduit.main.mixin;
+
+import org.spongepowered.asm.mixin.Mixins;
+import org.spongepowered.asm.mixin.connect.IMixinConnector;
+
+public class MixinConnector implements IMixinConnector {
+
+    @Override
+    public void connect() {
+        Mixins.addConfiguration("mixins.conduit.json");
+    }
+}
