+++ src/main/java/systems/conduit/main/mixins/ServerGamePacketListenerMixin.java
@@ -1,4 +1,4 @@
-package systems.conduit.main.mixin;
+package systems.conduit.main.mixins;
 
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.MinecraftServer;
