+++ src/main/java/systems/conduit/main/mixins/event/player/DamageMixin.java
@@ -1,4 +1,4 @@
-package systems.conduit.main.mixin.event.player;
+package systems.conduit.main.mixins.event.player;
 
 import com.mojang.authlib.GameProfile;
