+++ src/main/java/systems/conduit/main/mixins/event/player/ServerPlayerGameModeMixin.java
@@ -1,4 +1,4 @@
-package systems.conduit.main.mixin.event.player;
+package systems.conduit.main.mixins.event.player;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
